moved container styling logic and PlaylistCards to the container folder
